Change (simplify?) Mutation lifetimes
Previously, a Mutation was bound to the lifecycle of its SourceFile,
which made it a bit harder to for example iterate them.

Now, a Mutation takes a copy of the SourceFile, and this copy is made
cheaper by having it hold the source code in an Rc, so only a few paths
are copied.

So there's now a straightforward SourceTree::mutations method.

Also, add some rustdoc.