Refactor: remove Lab object
It seems like in Rust, aside from the Builder pattern, objects that are
built for a single-shot use often aren't a good fit. Perhaps if their
external interface only has a constructor and a single method to call
it's no better for the caller than a function?

And internally, it does give somewhat reduce the number of arguments to
pass to internal methods, but perhaps also makes it less clear which
ones are really needed. For example when we parallelize builds, the
assumption in the previous code that they'd all use the same build
directory would need to be removed, which makes the Lab even weaker.