Build the source tree before copying
The source tree target might be stale in various ways: rustup changed
the compiler, maybe it's just never been built, or maybe it hasn't built
some dependencies. The build in the scratch directory will generally
cope with all these, but it repeats that work every time cargo mutants
is run, which makes it slower to iterate during development.

So instead, we freshen the source directory before copying. If it's
already been recently built, this is very cheap. If not, it should not
add much time on the first mutants run, and will make others faster.