fix: `version` and `completions` cmd in non cargo directory
Sometimes `version` and `completions` subcommands are to run in a
directory that is not necessarily a Cargo project, for example, an IDE
plugin to check if `cargo-mutants` is installed or a shell script to
install completion for a dev env setup.

This commit allows both subcommands run inside a non Cargo project
directory.